+++ src/api/caidapp/views.py
@@ -539,7 +539,8 @@ class IdentityListView(LoginRequiredMixin, ListView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context["filter_form"] = self.filterset.form
+        # context["filter_form"] = self.filterset.form
+        context["filter"] = self.filterset
         context = add_querystring_to_context(self.request, context)
         # query_params = self.request.GET.copy()
         # query_params.pop('page', None)